pyproject.toml
[project]
name = "hello-pydanticai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langfuse>=3.0.2",
    "markdownify>=1.1.0",
    "pydantic-ai>=0.2.16",
    "pydantic-ai-slim[mcp]>=0.2.16",
    "readability-lxml>=0.8.4.1",
]
